moved facebook specific code to seperate django app
