+++ social_media_crawler/facebook_crawler/settings.py
@@ -1,0 +1,6 @@
+import facebook
+
+# Facebook configuration
+FACEBOOK_APP_ID = '377609379105449'
+FACEBOOK_SECRET_KEY = '058825d0ce4c2b507da708a09a180d20'
+
