+++ social_media_crawler/social_media_crawler/urls.py
@@ -3,6 +3,6 @@ from django.contrib import admin
 import views
 
 urlpatterns = patterns('',
-    url(r'^posts/(?P<link>.*)$', views.fetch_posts_for_link),
-    url(r'^admin/', include(admin.site.urls))
+    url(r'^admin/', include(admin.site.urls)),
+    url(r'^post/', include('facebook_crawler.urls'))
 )
